updated api version to 2.1.0 for BowTie Chart
