+++ src/CustomVisuals/Published/BowtieChart/pbiviz.json
@@ -4,12 +4,12 @@
     "displayName": "Bowtie Chart by MAQ Software",
     "guid": "BowtieChart1465021618681",
     "visualClassName": "BowtieChart",
-    "version": "3.1.3",
+    "version": "3.1.4",
     "description": "Bowtie chart",
     "supportUrl": "https://github.com/maqsoftware/PowerBI-visuals/issues",
     "gitHubUrl": "https://github.com/maqsoftware/PowerBI-visuals/"
   },
-  "apiVersion": "1.4.0",
+  "apiVersion": "2.1.0",
   "author": {
     "name": "MAQ Software",
     "email": "support@maqsoftware.com"
@@ -33,7 +33,7 @@
     "node_modules/powerbi-visuals-utils-interactivityutils/lib/index.js",
     "node_modules/powerbi-visuals-utils-chartutils/lib/index.js",
     "node_modules/powerbi-visuals-utils-colorutils/lib/index.js"
- ],
+  ],
   "style": "style/visual.less",
   "capabilities": "capabilities.json",
   "dependencies": "dependencies.json"
